+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v0.3.0
+
+* Introduce `WithNotifierBufferSize` and `WithEntrypointBufferSize` to `ReliableNotifier`
+* Change `ReliableNofifier` default buffer sizes from 0 to 10
+
 # v0.2.0
 
 * Introduce supervision restart strategy `OneForAll` (#65)